Adding RFM69 class
