+++ platformio.ini
@@ -40,14 +40,14 @@ lib_deps =
     ${env.lib_deps}
     Adafruit MCP9808 Library
     Wire
-src_filter = +<RFM69/MCP9808> -<RFM69/Gateway> -<RFM69/DHT11> -<LoRa/> 
+src_filter = +<RFM69/MCP9808> +<RFM69/> -<RFM69/Gateway> -<RFM69/DHT11> -<LoRa/> 
 
 [env:Uno_RFM69_DHT11]
 board = uno
 lib_deps = 
     ${env.lib_deps}
     SimpleDHT
-src_filter = +<RFM69/DHT11> -<RFM69/Gateway> -<RFM69/MCP9808> -<LoRa/> 
+src_filter = +<RFM69/DHT11> +<RFM69/> -<RFM69/Gateway> -<RFM69/MCP9808> -<LoRa/> 
 
 [env:Uno_RFM69_MCP9808]
 board = uno
@@ -55,10 +55,10 @@ lib_deps =
     ${env.lib_deps}
     Adafruit MCP9808 Library
     Wire
-src_filter = +<RFM69/MCP9808> -<RFM69/Gateway> -<RFM69/DHT11> -<LoRa/> 
+src_filter = +<RFM69/MCP9808> +<RFM69/> -<RFM69/Gateway> -<RFM69/DHT11> -<LoRa/> 
 
 [env:Uno_RFM69_Gateway]
 board = uno
 lib_deps = 
     ${env.lib_deps}
-src_filter = +<RFM69/Gateway> -<RFM69/MCP9808> -<RFM69/DHT11> -<LoRa/> 
+src_filter = +<RFM69/Gateway> +<RFM69/> -<RFM69/MCP9808> -<RFM69/DHT11> -<LoRa/>
