+++ src/RFM69/DHT11/DHT11.cpp
@@ -1,39 +1,40 @@
-#include <Wire.h>
-#include "Adafruit_MCP9808.h"
-#include "../RFM95.h"
+#include <SimpleDHT.h>
+#include "../RFM69.h"
+
+#define DHT11_PIN 4
 
 // Singleton instance of the radio driver
-RFM95 radio;
+RFM69 radio;
 
-// Create the MCP9808 temperature sensor object
-Adafruit_MCP9808 tempsensor = Adafruit_MCP9808();
+//Create DHT11 dht11 object
+SimpleDHT11 dht11;
 
 // Function prototypes
 int8_t celciusToFahrenheit(int8_t c);
 
 char data[20];
 char buffer[10];
+byte temp = -1;
+byte humidity = -1;
 
 void setup()
 {
-    radio.initRadio();
-
-    if (!tempsensor.begin())
-        Serial.println("Couldn't find MCP9808!");
+  radio.initRadio();
 }
 
 void loop()
 {
-    strcpy(data, "TEMPF:");
-    itoa(celciusToFahrenheit(tempsensor.readTempC()), buffer, 10);
-    strcat(data, buffer);
-    strcat(data, '\0');
-    Serial.println(data);
+  strcpy(data, "TEMPF:");
+  dht11.read(DHT11_PIN, &temp, &humidity, NULL);
+  itoa(celciusToFahrenheit(temp), buffer, 10);
+  strcat(data, buffer);
+  strcat(data, '\0');
+  Serial.println(data);
 
-    radio.sendMessage(data, 20);
+  radio.sendMessage(data, 20);
 
-    // Sends data every 5 seconds
-    delay(5000);
+  // Sends data every 5 seconds
+  delay(50000);
 }
 
 /**************************************************************
@@ -45,5 +46,5 @@ void loop()
 **************************************************************/
 int8_t celciusToFahrenheit(int8_t c)
 {
-    return (c * 9.0) / 5.0 + 32;
+  return (c * 9.0) / 5.0 + 32;
 }
